✨ Add CompanyCoffeeChatButton and CompanyCoffeeChatButtonGroup components for managing coffee chat selections
