+++ .env.dev
@@ -1,1 +1,1 @@
-VITE_API_BASE_URL = https://www.survey-josha.site
+VITE_API_BASE_URL = https://dev.survey-josha.site/
