+++ src/pages/CompanyMyPage.tsx
@@ -1,10 +1,5 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useQuery } from '@tanstack/react-query';
 import { useState } from 'react';
 
-import type { CoffeeChatStatus } from '@/api/apis/localServer/schemas';
-import { CoffeeChatButton } from '@/components/button/CoffeeChatButton';
-import { UpdateCoffeeChatStatusModal } from '@/components/modal/UpdateCoffeeChatStatusModal';
 import { GlobalNavigationBar } from '@/components/nav/GlobalNavigationBar';
 import { Button } from '@/components/ui/button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
@@ -14,15 +9,11 @@ import {
   CoffeeChatNumberBadge,
   CompanyCoffeeChatListView,
 } from '@/feature/coffeeChat';
-import { CompanyCoffeeChatBtnGroup } from '@/feature/coffeeChat/ui/mypage/company/CompanyCoffeeChatBtn';
+import { CompanyCoffeeChatButton } from '@/feature/coffeeChat/ui/mypage/company/CompanyCoffeeChatButton';
 import { CompanyProfileView } from '@/feature/company';
 import { MyPostList } from '@/feature/post/ui/mypage/company/MyPostList';
-import { useGuardContext } from '@/shared/context/hooks';
-import { ServiceContext } from '@/shared/context/ServiceContext';
-import { TokenContext } from '@/shared/context/TokenContext';
 import { useRouteNavigation } from '@/shared/route/useRouteNavigation';
 import { useRouteQueryParams } from '@/shared/route/useRouteParams';
-
 type MyPageTab = MyPageRouteQuery['tab'];
 
 export const CompanyMyPage = () => {
@@ -31,51 +22,11 @@ export const CompanyMyPage = () => {
   const [isExistProfile, setIsExistProfile] = useState(false);
   const [companyId, setCompanyId] = useState<string | null>(null);
   const [selectedChats, setSelectedChats] = useState<string[]>([]);
-  const { coffeeChatListData } = useGetCoffeeChatList();
-  const { updateCoffeeChatStatus, isPending } = useUpdateCoffeeChatStatus();
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [modalStatus, setModalStatus] = useState<'ACCEPTED' | 'REJECTED'>(
-    'ACCEPTED',
-  );
 
   const [isSelectMode, setIsSelectMode] = useState(false);
-  const handleConfirm = (status: CoffeeChatStatus) => {
-    updateCoffeeChatStatus({
-      coffeeChatList: selectedChats,
-      coffeeChatStatus: status,
-    });
-    setSelectedChats([]);
-    setIsModalOpen(false);
-  };
 
   const handleTabChange = (tab: string) => {
     toMyPage({ query: { tab: tab as MyPageTab } });
-  };
-  const handleOpenModal = (status: 'ACCEPTED' | 'REJECTED') => {
-    if (selectedChats.length === 0) {
-      alert('선택된 커피챗이 없습니다.');
-      return;
-    }
-    setModalStatus(status);
-    setIsModalOpen(true);
-  };
-
-  const handleSelectAll = () => {
-    if (
-      coffeeChatListData === undefined ||
-      coffeeChatListData.type === 'error'
-    ) {
-      return;
-    }
-    const allSelected = coffeeChatListData.data.coffeeChatList
-      .filter((coffeeChat) => coffeeChat.coffeeChatStatus === 'WAITING')
-      .map((coffeeChat) => coffeeChat.id);
-    setSelectedChats(allSelected);
-  };
-
-  const handleCancelSelect = () => {
-    setIsSelectMode(false);
   };
 
   return (
@@ -131,39 +82,12 @@ export const CompanyMyPage = () => {
                 )}
               </TabsContent>
               <TabsContent value="COFFEE_CHAT" className="ml-auto">
-                {coffeeChatListData?.type === 'success' &&
-                  coffeeChatListData.data.coffeeChatList.length > 0 &&
-                  (isSelectMode ? (
-                    <CompanyCoffeeChatBtnGroup
-                      isPending={isPending}
-                      handleOpenModal={handleOpenModal}
-                      handleCancelSelect={handleCancelSelect}
-                      handleSelectAll={handleSelectAll}
-                      disabled={isPending || selectedChats.length === 0}
-                    />
-                  ) : (
-                    <CoffeeChatButton
-                      onClick={() => {
-                        setIsSelectMode((prev) => !prev);
-                      }}
-                      disabled={isPending}
-                    >
-                      선택 하기
-                    </CoffeeChatButton>
-                  ))}
-
-                {isModalOpen && (
-                  <UpdateCoffeeChatStatusModal
-                    onClose={() => {
-                      setIsModalOpen(false);
-                    }}
-                    onConfirm={() => {
-                      handleConfirm(modalStatus);
-                    }}
-                    status={modalStatus}
-                    selectedCount={selectedChats.length}
-                  />
-                )}
+                <CompanyCoffeeChatButton
+                  isSelectMode={isSelectMode}
+                  setIsSelectMode={setIsSelectMode}
+                  selectedChats={selectedChats}
+                  setSelectedChats={setSelectedChats}
+                />
               </TabsContent>
             </TabsList>
 
@@ -188,76 +112,4 @@ export const CompanyMyPage = () => {
       </div>
     </div>
   );
-};
-
-const useGetCoffeeChatList = () => {
-  const { token } = useGuardContext(TokenContext);
-  const { coffeeChatService } = useGuardContext(ServiceContext);
-
-  const { data: coffeeChatListData } = useQuery({
-    queryKey: ['coffeeChatService', 'getCoffeeChatList', token] as const,
-    queryFn: ({ queryKey: [, , t] }) => {
-      if (t === null) {
-        throw new Error('토큰이 존재하지 않습니다.');
-      }
-      return coffeeChatService.getCoffeeChatList({ token: t });
-    },
-    enabled: token !== null,
-  });
-
-  if (coffeeChatListData?.type === 'success') {
-    coffeeChatListData.data.coffeeChatList.sort((a, b) => {
-      if (
-        a.coffeeChatStatus === 'WAITING' &&
-        b.coffeeChatStatus !== 'WAITING'
-      ) {
-        return -1;
-      }
-      if (
-        a.coffeeChatStatus !== 'WAITING' &&
-        b.coffeeChatStatus === 'WAITING'
-      ) {
-        return 1;
-      }
-      return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
-    });
-  }
-  return { coffeeChatListData };
-};
-
-const useUpdateCoffeeChatStatus = () => {
-  const { coffeeChatService } = useGuardContext(ServiceContext);
-  const { token } = useGuardContext(TokenContext);
-  const queryClient = useQueryClient();
-
-  const { mutate: updateCoffeeChatStatus, isPending } = useMutation({
-    mutationFn: ({
-      coffeeChatList,
-      coffeeChatStatus,
-    }: {
-      coffeeChatList: string[];
-      coffeeChatStatus: CoffeeChatStatus;
-    }) => {
-      if (token === null) {
-        throw new Error('토큰이 존재하지 않습니다.');
-      }
-      return coffeeChatService.updateCoffeeChatStatus({
-        token,
-        body: {
-          coffeeChatStatus,
-          coffeeChatList,
-        },
-      });
-    },
-    onSuccess: async (response) => {
-      if (response.type === 'success') {
-        await queryClient.invalidateQueries();
-      }
-    },
-  });
-
-  return {
-    updateCoffeeChatStatus,
-    isPending,
-  };
 };
